api/requirements.txt
fastapi==0.110.1
uvicorn[standard]==0.29.0
pydantic==2.6.4
pydantic-settings==2.2.1
httpx==0.27.0
redis==5.0.4
celery==5.3.6
python-multipart==0.0.9
prometheus-client==0.20.0
prometheus-fastapi-instrumentator==6.1.0
opentelemetry-sdk==1.23.0
opentelemetry-exporter-otlp==1.23.0
opentelemetry-instrumentation-fastapi==0.44b0
orjson==3.10.0
minio==7.2.4
tenacity==8.2.3
structlog==24.1.0
uvloop==0.19.0
asyncpg==0.29.0
